chore: remove redundant type cast (like using (float)1.0f)

+++ Prowl.Editor/Editor/NodeEditor/NodeEditor.cs
@@ -904,7 +904,7 @@ namespace Prowl.Editor
                                 draw++;
                                 if (draw >= 2) draw = -2;
                                 if (draw < 0) continue;
-                                if (draw == 0) bezierPrevious = CalculateBezierPoint(point_a, tangent_a, tangent_b, point_b, (j - 1f) / (float)division);
+                                if (draw == 0) bezierPrevious = CalculateBezierPoint(point_a, tangent_a, tangent_b, point_b, (j - 1f) / division);
                             }
                             Vector2 bezierNext = CalculateBezierPoint(point_a, tangent_a, tangent_b, point_b, j / (float)division);
                             gui.Draw2D.DrawLine(bezierPrevious, bezierNext, color, thickness);
