+++ Prowl.Runtime/Math/Vector2Int.cs
@@ -153,7 +153,7 @@ namespace Prowl.Runtime
 
             double ls = dx * dx + dy * dy;
 
-            return (double)Math.Sqrt((double)ls);
+            return Math.Sqrt(ls);
         }
 
         /// <summary>
